refactor text and texture aspects.
optimisation: texture sent to the GPU once after shader load
